Replace JSON parser with a YAML parser
The basic idea is to allow developers to
write YAML and JSON module-info files.

TO achieve that, pyyaml is used which for
some inexplicable reason doesn't support tabs.

The solution is to replace all tabs with spaces.
That way, the parser acts the same way as the original
json parser.

I'm not yet willing to go down the rabbit hole to see
the pitfalls of using YAML over JSON and why the hell
it doesn't support tabs, and what the possible impact
of doing a search and replace of all tabs.

I will just say that, as far as I am concerned, a
configuration file parsed by ego and epro will be
refactored by me to conform to the following:

1. internally, all tabs will be converted to spaces for
the yaml parser to be happy. This is simple plumbing.
2. under no circumstances should the original user file be
overwritten by a file that has all tabs replaced by spaces without user input. Doing yaml.dump() is not OK on a user file.
3. If supporting YAML becomes untenable due to security issues, I'll
drop support and refactor to use JSON again and a sane(r) configuration
file format.